# editoast/openapi_legacy.yaml
openapi: 3.0.2
info:
  title: OSRD Editoast
  description: OSRD Edition service description
  version: 0.1.0
...
paths:
...
  /rolling_stock/{id}/:
    get:
      tags:
        - rolling_stock
      summary: Retrieve a rolling stock
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: Rolling Stock ID
          required: true
      responses:
        200:
          description: The rolling stock information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RollingStock"
    patch:
      tags:
        - rolling_stock
      summary: Patch a rolling stock
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: Rolling Stock ID
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RollingStockUpsertPayload"
      responses:
        200:
          description: The updated rolling stock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RollingStock"
    delete:
      tags:
        - rolling_stock
      summary: Delete a rolling_stock and all entities linked to it
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: rolling_stock id
          required: true
        - in: query
          name: force
          description: force the deletion even if it’s used
          schema:
            type: boolean
            default: false
      responses:
        204:
          description: No content
        409:
          description: Rolling stock is used in a train schedule.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RollingStockUsage"
...
  /timetable/{id}/:
    get:
      tags:
        - timetable
      summary: Retrieve a timetable and its train schedules
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: Timetable ID
          required: true
      responses:
        200:
          description: The timetable content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableWithSchedulesDetails"
    post:
      description: Import a timetable
      parameters:
      - description: Timetable id
        in: path
        name: id
        required: true
        schema:
          format: int64
          minimum: 0
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/TimetableImportItem'
              type: array
        description: ''
        required: true
      responses:
        '204':
          description: Timetable was successfully imported
      summary: Import a timetable
      tags:
      - timetable
  /timetable/{id}/conflicts/:
    get:
      tags:
        - timetable
      summary: Retrieve all conflicts for a specific timetable
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: Timetable ID
          required: true
      responses:
        200:
          description: The timetable conflicts content
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    train_ids:
                      type: array
                      items:
                        type: number
                    train_names:
                      type: array
                      items:
                        type: string
                    start_time:
                      type: string
                    end_time:
                      type: string
                    conflict_type:
                      type: string
                  required:
                    [
                      train_ids,
                      train_names,
                      start_time,
                      end_time,
                      conflict_type,
                    ]
...
components:
  schemas:
    Battery:
      description: energy source for a rolling stock representing a battery
      type: object
      required:
        [
          energy_source_type,
          max_input_power,
          max_output_power,
          energy_storage,
          efficiency,
        ]
      properties:
        energy_source_type:
          type: string
          enum: ["Battery"]
        max_input_power:
          $ref: "#/components/schemas/SpeedDependantPower"
        max_output_power:
          $ref: "#/components/schemas/SpeedDependantPower"
        energy_storage:
          $ref: "#/components/schemas/EnergyStorage"
        efficiency:
          type: number
          format: float
          minimum: 0
          maximum: 1
    Catenary:
      description: energy source for a rolling stock representing a catenary
      type: object
      required:
        [energy_source_type, max_input_power, max_output_power, efficiency]
      properties:
        energy_source_type:
          type: string
          enum: ["Catenary"]
        max_input_power:
          $ref: "#/components/schemas/SpeedDependantPower"
        max_output_power:
          $ref: "#/components/schemas/SpeedDependantPower"
        efficiency:
          type: number
          format: float
          minimum: 0
          maximum: 1
    Comfort:
      description: Train comfort that will be used for the simulation
      type: string
      enum: ["AC", "HEATING", "STANDARD"]
      default: "STANDARD"
    ConditionalEffortCurve:
      type: object
      properties:
        cond:
          type: object
          nullable: true
          properties:
            comfort:
              $ref: "#/components/schemas/Comfort"
              nullable: true
            electrical_profile_level:
              type: string
              nullable: true
            power_restriction_code:
              type: string
              nullable: true
        curve:
          $ref: "#/components/schemas/EffortCurve"
    EffortCurve:
      type: object
      properties:
        speeds:
          type: array
          example: [0.0, 2.958, 46.719]
          minItems: 2
          items:
            type: number
            format: float
        max_efforts:
          type: array
          example: [23500.0, 23200.0, 21200.0]
          minItems: 2
          items:
            type: number
            format: float
    EnergySource:
      description: enery source of a rolling stock
      oneOf:
        - $ref: "#/components/schemas/Catenary"
        - $ref: "#/components/schemas/PowerPack"
        - $ref: "#/components/schemas/Battery"
      discriminator:
        propertyName: energy_source_type
...
    RollingStock:
      allOf:
        - $ref: "#/components/schemas/RollingStockUpsertPayload"
        - type: object
          properties:
            railjson_version:
              type: string
              example: "3.2"
            id:
              type: number
            liveries:
              type: array
              items:
                $ref: "#/components/schemas/RollingStockLivery"
          required: [railjson_version, id, liveries]
    RollingStockUpsertPayload:
      allOf:
        - $ref: "#/components/schemas/RollingStockBase"
        - type: object
          properties:
            effort_curves:
              type: object
              properties:
                default_mode:
                  type: string
                modes:
                  type: object
                  additionalProperties:
                    type: object
                    properties:
                      curves:
                        type: array
                        items:
                          $ref: "#/components/schemas/ConditionalEffortCurve"
                      default_curve:
                        $ref: "#/components/schemas/EffortCurve"
                      is_electric:
                        type: boolean
                    required: [curves, default_curve, is_electric]
              required: [default_mode, modes]
          required: [effort_curves]
    RollingStockBase:
      properties:
        name:
          type: string
        locked:
          type: boolean
          default: false
          description: Whether the rolling stock can be edited/deleted or not.
        length:
          type: number
          format: float
        max_speed:
          type: number
          format: float
        startup_time:
          type: number
          format: float
        startup_acceleration:
          type: number
          format: float
        comfort_acceleration:
          type: number
          format: float
        gamma:
          type: object
          properties:
            type:
              type: string
              enum: ["CONST", "MAX"]
            value:
              type: number
              format: float
          required: [type, value]
        inertia_coefficient:
          type: number
          format: float
        features:
          type: array
          items:
            type: string
        mass:
          type: number
          format: float
        rolling_resistance:
          type: object
          properties:
            A:
              type: number
            B:
              type: number
            C:
              type: number
            type:
              type: string
              enum: ["davis"]
          required: [A, B, C, type]
        loading_gauge:
          type: string
          enum:
            - "G1"
            - "G2"
            - "GA"
            - "GB"
            - "GB1"
            - "GC"
            - "FR3.3"
            - "FR3.3/GB/G2"
            - "GLOTT"
        base_power_class:
          type: string
          example: "5"
        power_restrictions:
          type: object
          description: Mapping of power restriction code to power class
          additionalProperties:
            type: string
          example:
            "C1US": "4"
            "C2US": "3"
        energy_sources:
          type: array
          items:
            $ref: "#/components/schemas/EnergySource"
        electrical_power_startup_time:
          type: number
          nullable: true
          format: float
          description: The time the train takes before actually using electrical power (in seconds). Is null if the train is not electric.
          example: 5.0
        raise_pantograph_time:
          type: number
          nullable: true
          format: float
          description: The time it takes to raise this train's pantograph in seconds. Is null if the train is not electric.
          example: 15.0
        metadata:
          type: object
          properties:
            detail:
              type: string
            family:
              type: string
            grouping:
              type: string
            number:
              type: string
            reference:
              type: string
            series:
              type: string
            subseries:
              type: string
            type:
              type: string
            unit:
              type: string
          example:
            detail: "BB 7200"
            family: "LOCOMOTIVES"
            grouping: "Locomotives électriques courant continu"
            number: "1"
            reference: "7200"
            series: "BB 7200"
            subseries: "GV"
            type: "Locomotives électriques"
            unit: "US"
          required:
            [
              detail,
              family,
              grouping,
              number,
              reference,
              series,
              subseries,
              type,
              unit,
            ]
      required:
        - name
        - length
        - max_speed
        - startup_time
        - startup_acceleration
        - comfort_acceleration
        - gamma
        - inertia_coefficient
        - features
        - mass
        - rolling_resistance
        - loading_gauge
        - base_power_class
        - power_restrictions
        - metadata
        - energy_sources
        - electrical_power_startup_time
        - raise_pantograph_time
    RollingStockLivery:
      properties:
        id:
          type: number
        name:
          type: string
        compound_image_id:
          type: number
          nullable: true
      required: [id, name, compound_image_id]
...
    TimetableWithSchedulesDetails:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        train_schedule_summaries:
          type: array
          items:
            $ref: "#/components/schemas/TrainScheduleSummary"
      required:
        - id
        - name
        - train_schedule_summaries
    TrainScheduleSummary:
      allOf:
        - $ref: "#/components/schemas/TrainSchedule"
        - type: object
          properties:
            arrival_time:
              type: number
            mechanical_energy_consumed:
              type: object
              properties:
                base:
                  type: number
                eco:
                  type: number
                  nullable: true
              required: [base, eco]
            stops_count:
              type: integer
            path_length:
              type: integer
            invalid_reasons:
              type: array
              items:
                $ref: "#/components/schemas/InvalidTrainValues"
          required:
            - arrival_time
            - mechanical_energy_consumed
            - stops_count
            - path_length
            - invalid_reasons
    InvalidTrainValues:
      type: string
      enum:
        - NewerRollingStock
        - NewerInfra
...
    RollingStockUsage:
      type: object
      properties:
        rolling_stock_id:
          type: integer
        usage:
          type: object
          properties:
            train_schedule_id:
              type: integer
            train_name:
              type: string
            project_id:
              type: integer
            project_name:
              type: string
            study_id:
              type: integer
            study_name:
              type: string
            scenario_id:
              type: integer
            scenario_name:
              type: string
          required:
            - train_schedule_id
            - train_name
            - project_id
            - project_name
            - study_id
            - study_name
            - scenario_id
            - scenario_name
      required:
        - rolling_stock_id
        - usage
...
    TimetableImportItem:
      properties:
        path:
          items:
            $ref: '#/components/schemas/TimetableImportPathStep'
          type: array
        rolling_stock:
          example: 2TGV2N2
          type: string
        trains:
          items:
            $ref: '#/components/schemas/TimetableImportTrain'
          type: array
      required:
      - rolling_stock
      - path
      - trains
      type: object
    TimetableImportPathSchedule:
      properties:
        arrival_time:
          format: date-time
          type: string
        departure_time:
          format: date-time
          type: string
      required:
      - arrival_time
      - departure_time
      type: object
    TimetableImportPathStep:
      properties:
        location:
          $ref: '#/components/schemas/TimetableImportPathLocation'
        schedule:
          additionalProperties:
            $ref: '#/components/schemas/TimetableImportPathSchedule'
          example:
            '7015':
              arrival_time: 2023-08-17T09:46:00+02:00
              departure_time: 2023-08-17T09:56:00+02:00
          type: object
      required:
      - location
      type: object
    TimetableImportPathLocation:
      discriminator:
        propertyName: type
      oneOf:
      - $ref: "#/components/schemas/TrackOffsetLocation"
      - $ref: "#/components/schemas/OperationalPointLocation"
    TimetableImportTrain:
      properties:
        departure_time:
          example: 2023-08-17T08:46:00+02:00
          format: date-time
          type: string
        name:
          example: '7015'
          type: string
      required:
      - name
      - departure_time
      type: object
...